mv ConvertEnvToStringSlice() to util.go
