+++ internal/task/util_test.go
@@ -21,3 +21,22 @@ func compareSlices(a, b []string) bool {
 	}
 	return true
 }
+
+func TestConvertEnvToStringSlice(t *testing.T) {
+	env := map[string]string{
+		"FOO": "bar",
+	}
+
+	expected := []string{"FOO=bar"}
+	actual := ConvertEnvToStringSlice(env)
+
+	if len(actual) != len(expected) {
+		t.Errorf("Expected %d, got %d", len(expected), len(actual))
+	}
+
+	for i, v := range actual {
+		if v != expected[i] {
+			t.Errorf("Expected %s, got %s", expected[i], v)
+		}
+	}
+}
